Made it a plugin
