+++ plugins/diff-highlight/index.html
@@ -1,0 +1,81 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+	<meta charset="utf-8" />
+	<link rel="icon" href="favicon.png" />
+	<title>Data-URI Highlight ▲ Prism plugins</title>
+	<base href="../.." />
+	<link rel="stylesheet" href="style.css" />
+	<link rel="stylesheet" href="themes/prism.css" data-noprefix />
+	<link rel="stylesheet" href="plugins/diff-highlight/prism-diff-highlight.css" data-noprefix />
+	<script src="scripts/prefixfree.min.js"></script>
+
+	<script>var _gaq = [['_setAccount', 'UA-33746269-1'], ['_trackPageview']];</script>
+	<script src="https://www.google-analytics.com/ga.js" async></script>
+</head>
+
+<body class="language-none">
+
+	<header>
+		<div class="intro" data-src="templates/header-plugins.html" data-type="text/html"></div>
+
+		<h2>Diff Highlight</h2>
+		<p>Highlights the code inside diff blocks.</p>
+	</header>
+
+	<section>
+		<h1>How to use</h1>
+
+		<p>Replace the <code>language-diff</code> of your code block with a <code>language-diff-xxxx</code> class to enable syntax highlighting for diff blocks.</p>
+
+		<p>Optional:<br>
+			You can add the <code>diff-highlight</code> class to your code block to indicate changes using the background color of a line rather than the color of the text.<br>
+			This is not compatible with the Funky theme.</p>
+	</section>
+
+	<section>
+		<h1>Example</h1>
+
+		<p>Using <code>class="language-diff"</code>:</p>
+
+		<pre><code class="language-diff"> function add(a, b) {
+-    return a - b;
++    return a + b;
+ }</code></pre>
+
+		<p>Using <code>class="language-diff diff-highlight"</code>:</p>
+
+		<pre><code class="language-diff diff-highlight"> function add(a, b) {
+-    return a - b;
++    return a + b;
+ }</code></pre>
+
+		<p>Using <code>class="language-diff-javascript"</code>:</p>
+
+		<pre><code class="language-diff-javascript"> function add(a, b) {
+-    return a - b;
++    return a + b;
+ }</code></pre>
+
+		<p>Using <code>class="language-diff-javascript diff-highlight"</code>:</p>
+
+		<pre><code class="language-diff-javascript diff-highlight"> function add(a, b) {
+-    return a - b;
++    return a + b;
+ }</code></pre>
+
+	</section>
+
+	<footer data-src="templates/footer.html" data-type="text/html"></footer>
+
+	<script src="prism.js"></script>
+	<script src="components/prism-diff.js"></script>
+	<script src="plugins/diff-highlight/prism-diff-highlight.js"></script>
+	<script src="scripts/utopia.js"></script>
+	<script src="components.js"></script>
+	<script src="scripts/code.js"></script>
+
+</body>
+
+</html>
